// src/lib-stock/models/StockLevel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StockLevel = {
    id?: string;
    productId?: string;
    productName?: string;
    productSku?: string;
    agencyId?: string;
    quantity?: number;
    availableQuantity?: number;
    lastUpdated?: string;
};